+++ packages/google/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/google
 
+## 2.0.14
+
+### Patch Changes
+
+- Updated dependencies [0294b58]
+  - @ai-sdk/provider-utils@3.0.9
+
 ## 2.0.13
 
 ### Patch Changes